Put Services.kt in clear files, so that at a glance people can see what public services there are.
Fix include

Correct doc links after code movement
