+++ node/src/main/kotlin/net/corda/node/services/messaging/ServiceRequestMessage.kt
@@ -3,8 +3,8 @@ package net.corda.node.services.messaging
 import com.google.common.util.concurrent.ListenableFuture
 import net.corda.core.messaging.MessageRecipients
 import net.corda.core.messaging.SingleMessageRecipient
-import net.corda.core.node.services.DEFAULT_SESSION_ID
 import net.corda.core.serialization.CordaSerializable
+import net.corda.node.services.api.DEFAULT_SESSION_ID
 
 /**
  * Abstract superclass for request messages sent to services which expect a reply.
