+++ node/src/test/kotlin/net/corda/node/messaging/InMemoryMessagingTests.kt
@@ -1,7 +1,7 @@
 package net.corda.node.messaging
 
-import net.corda.core.node.services.DEFAULT_SESSION_ID
 import net.corda.core.node.services.ServiceInfo
+import net.corda.node.services.api.DEFAULT_SESSION_ID
 import net.corda.node.services.messaging.Message
 import net.corda.node.services.messaging.TopicStringValidator
 import net.corda.node.services.messaging.createMessage
